requirements.txt
Flask
flask-cors
numpy
pandas
scikit-learn
Pillow
joblib
